Fix critical bugs in audio engine and gameplay view
This commit addresses multiple critical issues identified by code review:
memory leaks, unsafe array indexing, state inconsistencies,
and fragile ID generation.

Audio Engine Improvements:
- MetronomeAudioEngine: Prepare haptic generators at init
  to eliminate first-use latency on iOS
- MetronomeAudioEngine: Fix interruption handling
  - Ensure onInterruption?(false) ALWAYS called when ended
  - Attempt resume only if requested, log errors
  - Prevent state lockups from missed callbacks
- MetronomeAudioEngine: Fix NotificationCenter memory leak
  - Store observer tokens in properties (interruptionObserver, routeChangeObserver)
  - Explicitly remove observers in deinit using stored tokens
  - Prevents memory accumulation during gameplay

Gameplay Improvements:
- GameplayViewModel: Fix isPlaying state consistency
  - Move isPlaying = true AFTER track validation guard
  - Prevents isPlaying = true when no track available
  - Ensures all setup code only runs with valid track
- GameplayViewModel: Fix DrumBeat ID collisions
  - Change id from Int to UInt64 for unique identifiers
  - Implement monotonic counter (generateBeatId())
  - Replace fragile Int(timePosition * 1000) generation
  - Update beatToBeamGroupMap and cachedBeatPositions types
- GameplaySheetMusicView: Fix unsafe array indexing
  - Remove cachedBeatIndices indexing (crash risk)
  - Iterate over cachedDrumBeats directly
  - Eliminates bounds checking and array sync requirements
  - Better performance and cleaner code

Testing Improvements:
- TestEnvironment: Improve XCTest argument detection
  - Match flag form (-XCTestConfigurationFilePath)
  - Match key=value form (contains "XCTestConfigurationFilePath")
  - Handle all XCTest configuration formats
- GameplayViewModelTests: Update test expectations
  - testStartPlaybackWithoutTrack expects isPlaying = false
  - Correct behavior after guard fix

All fixes verified with:
- Build: Succeeded
- Tests: All passed (MetronomeEngineTests, GameplayViewModelTests, GameplayViewTests)
- SwiftLint: No new violations

Pre-existing SwiftLint violations remain (file length, type body length)
but no NEW violations were introduced by these changes.

+++ Virgo/views/DrumBeatView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 // MARK: - Supporting Models
 struct DrumBeat {
-    let id: Int
+    let id: UInt64
     let drums: [DrumType]
     let timePosition: Double
     let interval: NoteInterval
